Move shared bcftools helpers to separate file
