+++ docs/development/worker_metrics.md
@@ -100,7 +100,7 @@ The metrics are captured in the tasks themselves in `packages/divbase-api/src/di
 
 Metrics for wall time is captured with `time.time()`, and CPU time, average and peak RSS memory is captured with `psutil`. The Python process and the `bcftools` subprocesses are monitored separately. The Python process monitoring is initiated with `process = psutil.Process()`; CPU monitoring is started with `process.cpu_times()`; memory usage is collected by sampling RSS over the course of the process with `process.memory_info().rss` as explained in the [Memory Usage](#memory-usage-not-additive-across-processes) section above and this is implemented as a class named `MemoryMonitor` in `packages/divbase-api/src/divbase_api/worker/metrics.py`.
 
-The `bcftools` subprocesses CPU and memory monitoring is found in `packages/divbase-api/src/divbase_api/services/vcf_queries.py`. When the code spawns a subprocess, it uses `subprocess.Popen` (instead of the commonly used `subprocess.run`). The PID of each spawned process is collected and `psutil.Process(proc.pid)` is used to collect the data from each PID with the `psutil` methods `.cpu_times()` and `.memory_info().rss` just like for the Python process.
+The `bcftools` subprocesses CPU and memory monitoring is found in `packages/divbase-api/src/divbase_api/services/vcf_queries.py` (`BcftoolsQueryManager.run_current_command`). Subprocesses are spawned via `run_bcftools` in `services/bcftools_helpers.py`, which uses `subprocess.Popen` (instead of the commonly used `subprocess.run`). The PID of each spawned process is collected and `psutil.Process(proc.pid)` is used to collect the data from each PID with the `psutil` methods `.cpu_times()` and `.memory_info().rss` just like for the Python process.
 
 After a task completes (regardless of Celery task status, e.g. SUCCESS or FAILED), the metrics are stored in an in-memory cache structured as a dictionary of dictionaries. This is done with the helper function `_record_task_metrics` in `packages/divbase-api/src/divbase_api/worker/tasks.py`, which calls `store_task_metric()` to save each metric value along with a timestamp. The cached values are then used to update the respective Prometheus gauges via `update_prometheus_gauges_from_cache()` in `packages/divbase-api/src/divbase_api/worker/metrics.py`. The metrics are stored per task using the DivBase job ID (not the Celery task UUID!) and task name as labels. The metrics are exposed via the Prometheus metrics endpoint at `/metrics`.
